+++ src/Router.tsx
@@ -1,15 +1,15 @@
 import { createBrowserRouter } from "react-router-dom";
-import LandingPage from "./pages/LandingPage";
-import RootLayout from "./layouts/RootLayout";
-import SearchPage from "./pages/SearchResultsPage";
-import Constants from "./constants";
-import CommunityPage from "./pages/CommunityPage";
-import CreatePostPage from "./pages/CreatePostPage";
-import CreateCommunityPage from "./pages/CreateCommunityPage";
-import RegisterPage from "./pages/RegisterPage";
-import PostPage from "./pages/PostPage";
-import ExploreCommunitiesPage from "./pages/ExploreCommunitiesPage";
-import AnimatedPage from "./components/animate/AnimatedPage";
+import LandingPage from "@/pages/LandingPage";
+import RootLayout from "@/layouts/RootLayout";
+import SearchPage from "@/pages/SearchResultsPage";
+import CommunityPage from "@/pages/CommunityPage";
+import CreatePostPage from "@/pages/CreatePostPage";
+import CreateCommunityPage from "@/pages/CreateCommunityPage";
+import RegisterPage from "@/pages/RegisterPage";
+import PostPage from "@/pages/PostPage";
+import ExploreCommunitiesPage from "@/pages/ExploreCommunitiesPage";
+import AnimatedPage from "@/components/animate/AnimatedPage";
+import { Constants } from "@/lib/constants";
 
 const router = createBrowserRouter([
   {
