+++ src/api/auth.ts
@@ -1,13 +1,8 @@
-import { api } from "./index";
-import { UserDTO, UserRegisterRequestDTO } from "../types/dto";
+import { api } from "@/api/config";
+import { UserDTO, UserRegisterRequestDTO } from "@/types/dtos";
 
-async function register(
-  userRegisterRequestDTO: UserRegisterRequestDTO
-): Promise<boolean> {
-  const registerResponse = await api.post<UserDTO>(
-    "/auth/register",
-    userRegisterRequestDTO
-  );
+async function register(userRegisterRequestDTO: UserRegisterRequestDTO): Promise<boolean> {
+  const registerResponse = await api.post<UserDTO>("/auth/register", userRegisterRequestDTO);
 
   return registerResponse.status == 201;
 }
