+++ src/api/suggestions.ts
@@ -1,5 +1,5 @@
-import { api } from ".";
-import { PostSummaryDTO } from "../types/dto";
+import { api } from "@/api/confi";
+import { PostSummaryDTO } from "@/types/dtos";
 
 async function getSuggestedPosts(): Promise<PostSummaryDTO[]> {
   const response = await api.get<PostSummaryDTO[]>(`/suggestions/posts`);
