+++ src/layouts/RootLayout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from "../components/Header";
+import Header from "@/components/Header";
 
 function RootLayout() {
   return (
