+++ src/pages/SearchResultsPage.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from "react";
 import { useSearchParams } from "react-router-dom";
-import { search } from "../api/communities";
+import { search } from "@/api/communities";
 import { IconSearch } from "@tabler/icons-react";
-import PageTitle from "../components/PageTitle";
-import SearchResultSkeleton from "../components/skeletons/SearchResultSkeleton";
-import { CommunitySummaryDTO } from "@/types/dto";
-import SearchResultCommunityList from "../components/SearchResultCommunityList";
+import PageTitle from "@/components/PageTitle";
+import SearchResultSkeleton from "@/components/skeletons/SearchResultSkeleton";
+import { CommunitySummaryDTO } from "@/types/dtos";
+import SearchResultCommunityList from "@/components/SearchResultCommunityList";
 
 function SearchResultsPage() {
   // This hook grabs the search query params from your browser
