+++ src/types/index.ts
@@ -1,21 +1,3 @@
-type Community = {
-  name: string;
-  description: string;
-};
-
-type Post = {
-  id: string;
-  title: string;
-  community: Community;
-  body: string;
-};
-
-type Comment = {
-  id: string;
-  text: string;
-  comments: Comment[];
-};
-
 type VoteState = "UPVOTED" | "DOWNVOTED";
 
-export type { Community, Post, Comment, VoteState };
+export type { VoteState };
